requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-dotenv==1.0.1
google-generativeai==0.3.2
openai==1.12.0
httpx==0.26.0
aiohttp==3.9.3
backoff==2.2.1
pydantic==2.6.1